Add ssh_pxssh.py to brute force SSH passwords
It has some differences from the book.

* First, we do not use globals to keep the found state or number of
failures. Instead we communicate between the main thread and the
connection threads through a thread-safe queue. In fact we use two of
these structures: the first has size 1 and its purpose is to communicate
the found state. The second has size `MAX_FAILS` and its purpose is to
communicate to the main thread that a maximum number of failures has
been reached.
* We also handle some new exceptions that may arise.